Control panel refactoring
